Rename zinnia.templates to zinnia.templating to avoid warning with the templates directory alongside
